Making CI pass and renamind hp_repo.py for clarity while at it.
